+++ src/ui/index.ts
@@ -7,3 +7,4 @@ export { default as Modal } from './Modal';
 export { default as SectionTitle } from './SectionTitle';
 export { default as Select } from './Select';
 export { default as SegmentedControl } from './SegmentedControl';
+export { default as Switch } from './Switch';